# src/main/resources/static/docs/swagger.yaml
openapi: 3.0.1
info:
  title: Bank API
  description: API for a bank system
  contact:
    name: Mikhail Trifonov
    email: trifonov2812@gmail.com
  version: 0.0.1
tags:
  - name: Bank User Controller
    description: Operations with bank users
  - name: Auth Controller
    description: Operations with authentication
paths:
  /api/users/{userId}/contact-info/remove:
    delete:
      tags:
        - Bank User Controller
      summary: Delete some contact info for a user
      description: Allows to delete phone number(s) and/or email(s) for a user
      parameters:
        - in: path
          name: userId
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDeleteContactInfoRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankUserDTO'
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
  /api/users/{userId}/contact-info/add:
    patch:
      tags:
        - Bank User Controller
      summary: Add new contact info for a user
      description: Allows to add phone number(s) and/or email(s) for a user
      parameters:
        - in: path
          name: userId
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddDeleteContactInfoRequest'
      responses:
        200:
            description: OK
            content:
                application/json:
                  schema:
                    $ref: '#/components/schemas/BankUserDTO'
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
  /api/users/{userId}/contact-info/update:
    patch:
      tags:
        - Bank User Controller
      summary: Update contact info for a user
      description: Allows to update phone number(s) and/or email(s) for a user
      parameters:
        - in: path
          name: userId
          required: true
          description: User ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateContactInfoRequest'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankUserDTO'
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
  /api/users/search:
    get:
      tags:
        - Bank User Controller
      summary: Search for users
      description: Allows to search for users with the specified parameters
      parameters:
        - in: query
          name: birthDate
          schema:
            type: string
            format: date
          description: The birth date of the user to search for
        - in: query
          name: phone
          schema:
            type: string
          description: The phone number of the user to search for
        - in: query
          name: fullName
          schema:
            type: string
          description: The full name of the user to search for
        - in: query
          name: email
          schema:
            type: string
          description: The email of the user to search for
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          description: The page number to retrieve
        - in: query
          name: size
          schema:
            type: integer
            default: 10
          description: The number of users to retrieve per page
        - in: query
          name: sort
          schema:
            type: array
            items:
              type: string
            default: [ "id,asc" ]
          description: The sorting criteria
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/BankUserDTO'
                  pageable:
                    type: object
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
  /api/users/{userIdSender}/transfer-to/{userIdReceiver}/{amount}:
    post:
      tags:
        - Bank User Controller
      summary: Transfer money between users
      description: Allows to transfer money between users
      parameters:
        - in: path
          name: userIdSender
          required: true
          description: Id of the user sending the money
          schema:
            type: integer
            format: int64
        - in: path
          name: userIdReceiver
          required: true
          description: Id of the user receiving the money
          schema:
            type: integer
            format: int64
        - in: path
          name: amount
          required: true
          description: Amount of money to transfer
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankUserDTO'
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
  /api/auth/register:
    post:
      tags:
        - Auth Controller
      summary: Register a new user
      description: Allows to register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequest'
      responses:
        200:
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BankUserDTO'
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
  /api/auth/login:
    post:
      tags:
        - Auth Controller
      summary: Login a user
      description: Allows to login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        200:
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        400:
          description: Bad request, invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                badRequest:
                  value:
                    message: "Bad request, invalid input"
                    status: 400
        500:
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                internalServerError:
                  value:
                    message: "Internal server error"
                    status: 500
components:
  schemas:
    BankUserDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Id of the user
        fullName:
          type: string
          description: Full name of the user
        balance:
          type: integer
          format: int64
          description: Balance of the user
        birthDate:
          type: string
          format: date
          description: Birth date of the user
        phone:
          type: array
          items:
            type: string
          description: Phone numbers of the user
        email:
          type: array
          items:
            type: string
          description: Emails of the user
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Error message
        status:
          type: integer
          format: int32
          description: Error status code
      description: Error response DTO
    RegisterUserRequest:
      type: object
      properties:
        login:
          type: string
          description: Login of the user
        password:
          type: string
          description: Password of the user
        fullName:
          type: string
          description: Full name of the user
        birthDate:
          type: string
          format: date
          description: Birth date of the user
        initialBalance:
          type: number
          format: big decimal
          description: Initial balance of the user
        phone:
          type: array
          items:
            type: string
          description: Phone number of the user
        email:
          type: array
          items:
            type: string
          description: Email of the user
      required:
        - login
        - password
        - initialBalance
        - phone
        - email
    LoginRequest:
      type: object
      properties:
        username:
          type: string
          description: Username of the user
        password:
          type: string
          description: Password of the user
      required:
        - username
        - password
    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT token
        tokenType:
          type: string
          default: "Bearer "
          description: Type of the token
    AddDeleteContactInfoRequest:
      type: object
      properties:
        phoneNumbers:
          type: array
          items:
            type: string
          description: Phone numbers to add or delete
        emailAddresses:
          type: array
          items:
            type: string
          description: Emails to add or delete
    UpdateContactInfoRequest:
      type: object
      properties:
        phoneNumbers:
          type: object
          additionalProperties:
            type: string
          description: Phone numbers to update
          example:
              "+79999999999": "+72345678901"
              "+78888888888": "+79876543210"
        emailAddresses:
          type: object
          additionalProperties:
            type: string
          description: Emails to update
          example:
              "mailToChange1@example.com": "mail@example.com"
              "mailToChange2@example.com": "mail2@example.com"